+++ include/mp++/concepts.hpp
@@ -80,90 +80,15 @@ MPPP_CONCEPT_DECL cpp_arithmetic = is_cpp_arithmetic<T>::value;
 
 #endif
 
-// Type trait to check if T is a supported integral type.
-// NOTE: mppp::detail::is_integral, for consistency with std::is_integral, will be true also for cv qualified
-// integral types. is_cpp_integral_interoperable, however, is used in contexts where the cv qualifications
-// matter, and we want this type trait not to be satisfied by integral types which are, e.g.,
-// const qualified.
 template <typename T>
-using is_cpp_integral_interoperable
-    = detail::conjunction<detail::is_integral<T>, std::is_same<detail::remove_cv_t<T>, T>>;
+using is_cpp_complex = detail::disjunction<std::is_same<T, std::complex<float>>, std::is_same<T, std::complex<double>>,
+                                           std::is_same<T, std::complex<long double>>>;
 
-template <typename T>
 #if defined(MPPP_HAVE_CONCEPTS)
-MPPP_CONCEPT_DECL CppIntegralInteroperable = is_cpp_integral_interoperable<T>::value;
-#else
-using cpp_integral_interoperable_enabler = detail::enable_if_t<is_cpp_integral_interoperable<T>::value, int>;
-#endif
-
-// Type trait to check if T is a supported unsigned integral type.
-template <typename T>
-using is_cpp_unsigned_integral_interoperable
-    = detail::conjunction<is_cpp_integral_interoperable<T>, detail::is_unsigned<T>>;
 
 template <typename T>
-#if defined(MPPP_HAVE_CONCEPTS)
-MPPP_CONCEPT_DECL CppUnsignedIntegralInteroperable = is_cpp_unsigned_integral_interoperable<T>::value;
-#else
-using cpp_unsigned_integral_interoperable_enabler
-    = detail::enable_if_t<is_cpp_unsigned_integral_interoperable<T>::value, int>;
-#endif
+MPPP_CONCEPT_DECL cpp_complex = is_cpp_complex<T>::value;
 
-// Type trait to check if T is a supported signed integral type.
-template <typename T>
-using is_cpp_signed_integral_interoperable
-    = detail::conjunction<is_cpp_integral_interoperable<T>, detail::is_signed<T>>;
-
-template <typename T>
-#if defined(MPPP_HAVE_CONCEPTS)
-MPPP_CONCEPT_DECL CppSignedIntegralInteroperable = is_cpp_signed_integral_interoperable<T>::value;
-#else
-using cpp_signed_integral_interoperable_enabler
-    = detail::enable_if_t<is_cpp_signed_integral_interoperable<T>::value, int>;
-#endif
-
-// Type trait to check if T is a supported floating-point type.
-template <typename T>
-using is_cpp_floating_point_interoperable
-    = detail::conjunction<std::is_same<detail::remove_cv_t<T>, T>, std::is_floating_point<T>
-#if !defined(MPPP_WITH_MPFR)
-                          ,
-                          detail::negation<std::is_same<T, long double>>
-#endif
-                          >;
-
-template <typename T>
-#if defined(MPPP_HAVE_CONCEPTS)
-MPPP_CONCEPT_DECL CppFloatingPointInteroperable = is_cpp_floating_point_interoperable<T>::value;
-#else
-using cpp_floating_point_interoperable_enabler
-    = detail::enable_if_t<is_cpp_floating_point_interoperable<T>::value, int>;
-#endif
-
-template <typename T>
-using is_cpp_interoperable
-    = detail::disjunction<is_cpp_integral_interoperable<T>, is_cpp_floating_point_interoperable<T>>;
-
-template <typename T>
-#if defined(MPPP_HAVE_CONCEPTS)
-MPPP_CONCEPT_DECL CppInteroperable = is_cpp_interoperable<T>::value;
-#else
-using cpp_interoperable_enabler = detail::enable_if_t<is_cpp_interoperable<T>::value, int>;
-#endif
-
-template <typename T>
-using is_cpp_complex = detail::disjunction<std::is_same<T, std::complex<float>>, std::is_same<T, std::complex<double>>
-#if defined(MPPP_WITH_MPFR)
-                                           ,
-                                           std::is_same<T, std::complex<long double>>
-#endif
-                                           >;
-
-template <typename T>
-#if defined(MPPP_HAVE_CONCEPTS)
-MPPP_CONCEPT_DECL CppComplex = is_cpp_complex<T>::value;
-#else
-using cpp_complex_enabler = detail::enable_if_t<is_cpp_complex<T>::value, int>;
 #endif
 
 namespace detail
@@ -190,20 +115,13 @@ using is_string_type
 #endif
                           >;
 
-template <typename T>
-#if defined(MPPP_HAVE_CONCEPTS)
-MPPP_CONCEPT_DECL StringType = is_string_type<T>::value;
-#else
-using string_type_enabler = detail::enable_if_t<is_string_type<T>::value, int>;
-#endif
-
 #if defined(MPPP_HAVE_CONCEPTS)
 
-// TODO move up.
 template <typename T>
 MPPP_CONCEPT_DECL string_type = is_string_type<T>::value;
 
 #endif
+
 } // namespace mppp
 
 #endif